requirements.txt
allure-pytest==2.13.3
allure-python-commons==2.13.3
assertpy==1.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pytest==8.0.0
requests==2.32.0
urllib3==2.2.2
python-dotenv~=1.0.1